Changes made that were discussed in last meeting. Namely, added caching logic in a new IdsService that replaced old Cache class. History and Reference services now pass the global identifier to their respective repository classes so that logic has been removed from the history and reference repositories. Parameter validation was added to controller. New service called CacheHydratingService was added to initially populate the cache with information from IdsRepository.This service was injected into the build in program.cs. Finally, the Value fields in the History model and HistoryDataitem model were changed to float from double.

+++ ViewpointAPI/Models/HistoryDataItem.cs
@@ -7,6 +7,6 @@ namespace ViewpointAPI.Models
     public class HistoryDataItem
     {
         public DateTime Timestamp { get; set; }
-        public double Value { get; set; }
+        public float Value { get; set; }
     }
 }
